refactor: 이동 횟수 정수형 타입 변경

+++ src/main/java/racingcar/Race.java
@@ -34,7 +34,7 @@ public class Race {
 		return round;
 	}
 
-	public boolean isRaceFinished(Long roundLimit) {
+	public boolean isRaceFinished(Integer roundLimit) {
 		return round >= roundLimit;
 	}
